Use built-in `stream.Transform` to create `HashStream`s
Since we're only targeting newer Node versions with a reasonable stream module, and we can use the `class` syntax sugar, it makes sense to extend `stream.Transform` directly to create a `HashStream`. In a future commit, this will allows to remove `through2` altogether.

Also, `HashStream` was previously mistyped as a `stream$Readable`, when in fact, it's a `stream$Transform`.