Properly save scoped packages in offline mirror.
This resolves a bug in which yarn was saving scoped packages in the offline mirror by their base name, which meant that if two packages in a project had the same name and version, but one of them was scoped and the other was not, only one of these would be saved in the offline mirror, and this would potentially cause a "bad hash" error to occur.

This commit fixes this behavior, and saves scoped npm packages in the offline mirror in the form `@[scope]-[package name]-[version].tar.gz`.