Use auth tokens for NPM registry requests
The NPM registry and NPM compatible registries support privately hosted
modules. To access meta data for these modules and the modules' tarballs
themselves, bearer tokens can be used.

To support this, NPM's `_auth` property need to be raed from the config
and this option must be used for meta data and tarball requests as the
bearer token.