Add a newline after the global comment in the lockfile (#1192)
The lockfile is nicely visually separated except for the first package which is right after the comment. The comment is global, though, not related directly to the first package so in order to make it easier to visually navigate the file and to mark the comment as global and not just related to the first package, I propose a newline is added after the comment.