Make `.git` suffix optional for hosted git
When a github URL is used in a `package.json` file which has a `.git`
suffix attached to its repository name, resolution fails with an error:

```
TypeError: Cannot read property 'split' of undefined
 at Function.parseRefs (/usr/local/Cellar/yarn/0.16.0/libexec/lib/node_modules/yarn/lib/util/git.js:439:20)
 at /usr/local/Cellar/yarn/0.16.0/libexec/lib/node_modules/yarn/lib/resolvers/exotics/hosted-git-resolver.js:138:50
 ...
```

for e.g. `https://github.com/gulpjs/gulp.git`.

This was causes by an extraneous `.git` suffix appended to the URL by
the resolver.

Fixes #1287