Fix install crashing when only stderr is not a TTY
Check that stderr is a TTY before rendering the progress bar, which
requires a TTY specific `stderr.columns` property.

The code previously checked if *stdout* is a TTY, which fails when
stdout is a TTY, but stderr is not. Now we check the stream that we're
actually going to use.