Personal access token for github over http - via env GITHUB_HTTP_AUTH
This PR will support importing `github:org/repo` when they are private.

```
GITHUB_HTTP_AUTH=user:personal_access_token yarn install
```

To support importing dependencies hosted in a private github repo
over http, the user authentication needs to be passed:
https://developer.github.com/v3/auth/#basic-authentication

This PR is a proof of concept: it does that with a shamelessly simple env
variable.
If this makes sense we can consider doing it cleanly.