Ability to opt-out of offline mirror
Configs are merged with configs in parent directories, so currently if a .yarnrc in a parent directory enables the offline mirror, all projects in child directories would also use the offline mirror.

This commit allows a child directory to opt-out of the offline mirror in their own .yarnrc with `yarn-offline-mirror false`.

Previously if you specified `false` you'd get an exception as false is parsed as boolean but it expects a string.

Also included tests verifying the existing .yarnrc merging behaviour (previously wasn't covered).

t14078443