[breaking?] `--check-files` for install/add/upgrade/check commands that verifies node_modules consistency (#3025)
Summary

This PR expands internal implementation of integrity check functionality.

.yarn-integrity file now contains all the properties as exploded JSON instead of sha-1 hashed.
This makes it easier to debug issues with integrity checks and installations

.yarn-integrity now contains list of files in node_modules if install/add command has --check-files flag passed.
When check command is executing it will have an option to test that files exist on disk.
This makes the command useful to identify when node_modules gets partially removed.

To use the new file checking feature:

yarn install --check-files
yarn check --integrity --check-files
This is an opt-in feature so people should not be affected buy it yet.
In the next iteration we may consider making --check-files the default option.
Also we may want to reevaluate other flags in check command and end up with more semantic/self explanatory options.

Exploded .yarn-integrity and especially files field allows us to integrate with third-party tools that may need list of files before yarn install is executed (e.g. BUCK).
Also it allows us to make a quick integrity check without loading much of the Yarn internals - only check that lock files entries are present, flags are the same and folders exist.