Allow emojis to be force enabled with `--emoji`.
The previous behaviour had a default of enabling emojis for Macs, and
allowed explicitly disabling them.

This PR keeps that default but allows `--emoji` to enable for other
platforms if the user wants to.  `--no-emoji` still works as before, but
due to the way that commander processes "--no-" flags (forcing the
default to true and overwriting the default value given in the `.option`
definition) it was necessary to switch the flag definition from
`--no-emoji` to `--emoji`.