Include yarn and node versions in yarn.lock
Sometimes it is helpful for automated tooling to know the version of
yarn and nodejs that was used in the environment that last modified the
yarn.lock file.

For example, this allows CI platforms to automatically ensure the
same versions of yarn and node are used in CI as locally.

This functionality can be disabled by adding to .yarnrc
  yarn-disable-lockfile-versions true