Tighten up upper bounds so that "2.0.0-pre" doesn't match "^1.0.0"
node-semver converts ~ and ^ ranges into pairs of >= and < ranges but the upper bounds don't properly exclude prerelease versions. For example, "^1.0.0" is converted to ">=1.0.0 <2.0.0", which includes "2.0.0-pre" since prerelease versions are lower than their non-prerelease counterparts. As a practical workaround we make upper-bound ranges exclude prereleases and convert "<2.0.0" to "<2.0.0-0", for example.

Added unit tests as well.