Give higher priority to .npmignore over .gitignore
If .npmignore or .yarnignore are present in a directory that also
contains .gitignore, don't take into account .gitignore when running the
pack command.

New function filterOverridenGitignores takes a WalkFiles object with all
the potential package files and returns another WalkFiles object only
with the relevant dot ignore files (.npmignore, .yarnignore or
.gitignore) taking into account the priority rules. This new collection
is the used by the pack command to create the ignore regexes.