Fix: Send auth token only when forced or mutating (fixes #1952)
Currently, when `this.token` is set, the auth token is always sent,
even with read-only GET and HEAD requests. This seems to be causing
yarn registry to error out. This patch changes the behavior so the
token is only sent with mutating HTTP requests (such as `PUT`) or
when it is forced via `always-auth`.