Fix: Use a dedicated sink for tar reading instead of a transform stream
**Summary**

Potentially fixes #3011 (specifically https://git.io/vHAzW)

Turns out our `ConcatStream` implementation was only used to read
from `tar-fs` when creating a package to upload and put that stream
into a memory buffer. Since `ConcatStream` was implemented as a
`stream.Transform` and nothing was reading back from it, it had the
potential to just hang there until something reads from it. This
patch replaces that with [a small script][1].

[1]: http://www.geekpeak.de/images/produkte/i22/22-go-away-or-i-will-replace-you-de.jpg

**Test plan**

Removed existing tests for the old module. Rely on other existing
tests for the replacement code.