Fix: inherit env variables from process when calling Git
**Summary**

Fixes #3742.

With #3633 we started overriding the default `env` map that was passed
to `git` calls. It defaulted to `process.env` and since the option is
a whitelist, now it is only the ones we wanted to add, instead of
`process.env` + the additions. This patch fixes the problem by spreading
`process.env` before the new values.

**Test plan**

Manual verification for t19705486. We should have automated tests for this
so I'll start working on them. Just want the fix our first.