Make leading `.` part of the config location.
Moved the leading `.` from the config filename to be added by the location searcher.  The alternative is to strip the leading `.` to create the global name, but that feels clumsy.  This should also make it easier to make portable if the leading `.` is not used for hidden config files on other platforms.

Added some comments explaining the path transformations.