Breaking: rename every `yarn <command> rm` to `yarn <command> remove` (#3989)
**Summary**

From #3952 - For the sake of consistency and clarity, we need to rename the `rm` command to `remove` and add a deprecation warning for `rm`.

**Test plan**

Only manual testing for now. Unfortunately, every command with `rm` requires a login. Coincidentally, and if I am not mistaken, none of the commands that implement `rm` has tests ( `owner`,  `tag`, `team`).