Fix: Only add auth token when registry url matches pathname url (#3987)
**Summary**
Fixes: #3907.

In yarn version 0.27.4 a patch (https://github.com/yarnpkg/yarn/commit/5ff69228f6cc9fa27bbaeb67412f7f66f218b9da) introduced a change that caused an authorization header to be added to requests that were not being sent to the registry.

This commit brings back some older logic to ensure we only add the authorization header when we are sending requests to the registry.

**Test plan**

Many new unit tests for requests and `isRequestToRegistry`.

To test this change you need a private repository using a repository server such, as Verdaccio configured, to only allow authenticated calls to be made for both download and publish commands. This configuration requires that yarn has the `always-auth` option enabled.

If we install additional packages from GitHub and save them to the `yarn.lock` file it should not send the authorization header for these requests.