Fix: handle potential race conditions when using cache (#3985)
**Summary**

There's a delay between when we make sure a file in cache exists and when we actually start reading it. Our code wasn't doing any error handling when trying to open the file so if another process removes the file we assume existing we were throwing an unexpected `ENOENT ... open` error.

This patch removes all delayed existence checks and catches the error directly on the `ReadStream` instance directly. Moreover, it automatically falls back for local caches: direct, cache, offline mirror.

This revealed a bug in tarball path resolution which I fixed.

**Test plan**

Existing tests.