Update: Add support for Github URL shorthand to yarn init (#4012)
**Summary**
This pull request fulfills #4006 – an issue that requests a feature be added to `yarn init` so that Github shorthand i.e. a value of `yarnpkg/yarn` can be used in the repository field.

**Test plan**
I have added a test to the test suite. Outside of that, this is the difference in usage:

**Before:** Running `yarn init` with default options other than the repository field which is set to `yarnpkg/yarn` would yield a package.json that has the repository URL set to `yarnpkg/yarn` 

**After:** Running `yarn init` with default options other than the repository field which is set to `yarnpkg/yarn` would yield a package.json that has the repository URL set to `https://github.com/yarnpkg/yarn`