Fix: Fix numeric versions resolving to local files
**Summary**

Fixes #4013. We were blindly priotizing local files (even if they
are not directories) if there's anything on the file system matching
a pattern for a package. This patch ensures that we only do this if
the pattern starts with `./`, `../`. `/`, `~/` (and their `\` versions
for Windows) and the matched local entity is a directory. The former
part is in line with [NPM specs](https://docs.npmjs.com/files/package.json#local-paths).

**Test plan**

Added a new unit test and modified an incorrect old one. Also manually
verified that the issue described in #4013 does not happen anymore.