Fix: deadlock when using network mutex and concurrent runs
**Summary**

The current network mutex code relies on the old `process.exit()` to
terminate all existing connections. Since that has been removed now,
when a "master" `yarn` instance gets another one connected to it and
waiting for it to end, it stays in a zombie state due to the TCP
server not terminating with clients on it. This patch makes it shut
the server down properly when everything is done, fixing the deadlock.

**Test plan**

Adds a new integration test which hangs and times out without the fix.
Also can be tested manually, running two `yarn` instances side by side
with the `--mutex=network` option, concurrently.