Fix: Clean-up cwd after integration tests
**Summary**

We normally clean up the test `cwd` after running them but we
don't for integration tests. With the addition of a new test case,
this started causing macOS builds to fail on TravisCI due to exhaustion
of space on ramdisk which we use to run tests on for speed.

**Test plan**

macOS tests should pass on TravisCI.