Fix: Clean-up cwd after integration tests and increase ramfs size (#4108)
**Summary**

We normally clean up the test `cwd` after running them but we
don't for integration tests. With the addition of a new test case,
this started causing macOS builds to fail on TravisCI due to exhaustion
of space on ramdisk which we use to run tests on for speed. This patch
also increases ramfs size on Travis to make tests pass since clean up
alone was not enough.

**Test plan**

macOS tests should pass on TravisCI.