Fix: stop all active spinners on reporter.close()
**Summary**

We were not stopping any activities in the reporter and were
relying on `process.exit()` to kill them earlier. This patch
fixes that and ensures all acitivies are stopped when the
reporter is closed.

**Test plan**

Adds 3 new snapshots tests that fail without the fix.

Manual: just try to install a non-existent package and see `yarn`
getting stuck due to an ongoing reporter activity. After the fix,
it shuts down properly.