Fix: Reimplements part of the integrity check
**Summary**

The integrity file doesn't currently support workspaces, which leads
to #3863 (removing a workspace node_modules folder and running yarn
doesn't do anything, because Yarn incorrectly assumes that everything
is already done).

This PR rewrites a good chunk of the integrity logic to support and
fix this behavior.

**Test plan**

Tests are passing, I still have to write an automated test before
merging this PR.