Fix: invalid linking in workspaces for unused packages (#4129)
**Summary**

Fixes #4106. When a local package is listed under `workspaces` but that specific version is never used by anything, `yarn` still tries to install its bin links because it is listed as a dependency of the virtual workspace aggregator (but never get installed). This patch makes sure we have a location inside `node_modules` when trying to create bin links. It also fixes where we don't pass the proper location for workspaces root which may result in unexpected behavior under certain circumstances.

**Test plan**

Added a new test case which fails on master and passes with the fix.

Manual:
```
git clone git@github.com:babel/babel.git
yarn
```

Fails before the patch with latest nightly and passes after the patch.