Fix: stop all active spinners on reporter.close() (#4150)
**Summary**

We were not stopping any activities in the reporter and were
relying on `process.exit()` to kill them earlier. This patch
fixes that and ensures all activities are stopped when the
reporter is closed.

**Test plan**

Adds a new snapshots test that fails without the fix.

Manual: just try to install a non-existent package and see `yarn`
getting stuck due to an ongoing reporter activity. After the fix,
it shuts down properly.