Update: Prevent `yarn add` from being run from the workspace root directory (#4166)
**Summary**

This PR prevents using `yarn add <pkg>` from within a workspace root folder, in order to prevent common mistakes where a package is added to the workspace root folder instead of being added to the right workspace (which, because of the node resolution algorithm, would be effectively invisible to the user).

If the user really wants to add a dependency to the workspace root folder, then we suggest them to use `yarn add <pkg> --dev` instead. This should be fine since the workspaces will probably only be used in development mode anyway. An alternative would be to introduce a new flag `--force` that would bypass this check, but I think the current heuristic is fine until we receive complaints.

**Test plan**

Added two tests.