Fix: Use semver tag to colorize package name for outdated/upgrade-interactive commands (#4183)
**Summary**
This will fix #3815 by implementing a solution mentioned here https://github.com/yarnpkg/yarn/issues/3815#issuecomment-322496710.
As the `outdated`-command displayed pretty much the same data I changed the colors of that command as well.

New output: 
![screen shot 2017-08-16 at 10 22 05](https://user-images.githubusercontent.com/2036823/29357714-c91f457a-8278-11e7-99fc-ef1a3be65b3e.png)
![screen shot 2017-08-16 at 11 48 54](https://user-images.githubusercontent.com/2036823/29357746-e4c1c1ea-8278-11e7-8541-6f71213d686d.png)

**Test plan**

I did not create any new tests as this is only a visual change and was not tested before. But if we decide to move the new "helper files" to a utils folder I can write a few tests to test the output.