Fix: Allow resolver to delay resolutions for nested dependencies (#4372)
**Summary**
Fix for https://github.com/yarnpkg/yarn/issues/4357.

The issue is that when a dependency (dep A) defined in resolutions (includes dep A, dep B) depends on another resolution (dep B), then it expects to match its own nested dep B to the top level resolution dep B. So the first part of this fix is "don't run resolutions map check when it's in flat mode", which is what threw the invariant warning. Second part of the fix is that we still want that nested dependency (dep B) of a resolution (dep A) to be resolved correctly. `--flat` mode solves this by [collapsing all versions](https://github.com/yarnpkg/yarn/blob/master/src/cli/commands/install.js#L558) after the resolver is done. For resolutions, I'm adding a delay queue for requests with resolutions but no manifests found yet so that they will be resolved later.

**Test plan**
Added tests in resolutions