Fix: Diffing identical versions should not include extra period at end (#4428)
**Summary**
Fixes #4395 

This PR fixes a bug that causes `colorizedDiff` to append a period at the end of the return value if the `from`/`to` versions are identical. If they do happen to be identical, then it will simply return the `from` value.

**Test plan**
I am currently testing that, given identical `from`/`to` versions, it returns the `from` (or `to`) version. I've not added unit tests to cover when the versions differ in any way, nor am I testing that the returned string is formatted correctly (e.g., if `from = '1.0.0'` and `to = '1.0.1'`, the result should be `1.0.1` with the trailing one being formatted to green)