Fix: Handle `null` versions in package.json (#4461)
**Summary**

Fixes #4429. Apparently there's an advice on the internet to use
`null` as the package version to install the latest version and
even worse, some people are using this. This patch makes sure
the version is at least a string (albeit an empty one) when
fixing the manifest.

**Test plan**

Added new integration test.