fix(network): Allow `https-proxy false` to not use a proxy for HTTPS (#4614)
**Summary**

Fixes #4546. Previous behavior was that if `.npmrc` or `.yarnrc` contained and `http-proxy` but not an `https-proxy`, HTTPS requests would "fall back" to the `http-proxy`.

This is not always the desired behavior; See #4546 sometimes one protocol needs a proxy and the other does not.

This PR adds the ability to set the `https-proxy` to `false` which will cause Yarn to not use an HTTPS proxy (and not fall-back to the HTTP proxy). This allows the code to treat `undefined` as the setting not being specified (which will fall back to http-proxy) and `false` as the setting being specified, but set to false (which will not use a proxy).

**Test plan**

Tested manually on Windows with Fiddler.

I have no idea how to test this since requests go through the mock http requester when running tests. Suggestions welcome!