fix(request-manager): Flaky network messsage should be info, not warning (#4621)
**Summary**

Sometimes Yarn saturates the network and sometimes the network itself
has issues, especially on CI and Yarn reports this as a warning. This
should be a simple info message unless it actually completely fails.

This also fixes a flaky integration test with react-scripts.

**Test plan**

`react-scripts` integration test should not be flaky anymore.