chore(tests): Don't output coverage on regular test runs (#4629)
* chore(tests): Don't output coverage on regular test runs

**Summary**

We output a long coverage summary in both CI and local tests which
just adds noise to the output, especially when tests are failing
and one is trying to figure out what is wrong. Since this
information is not used by any automatic tooling, it is better to
disable it for now. This PR keeps the option with `test-coverage`
script for people who still want it.

**Test plan**

When running `yarn test` you should not see a long coverage table
at the end after the patch.

* some fixes

* Revert "some fixes"

This reverts commit 75ca538995c606056e3358c794caa3c19a09ee10.