quickwin for failed fetch of missing fsevents
fsevents is an optional dependency. Although platforms such as linux ignore the failed fetch, I'm having a situation that Yarn still wants to resolve the version due to conflicting consumer plugins and eventually adds it as a resolution to package.json (and to the lock file).
If you want to re-run yarn install, Yarn insists on checking the fsevents entry at "resolutions" without having the information that it's just optional.
My approach is a bit more generic that just "is it fsevents" and simply checks whether the config holds a location for the current item. If falsy, pretend that it's optional.